QuantLab/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
ccxt>=4.0.0
requests>=2.31.0

sqlalchemy>=2.0.0
python-dotenv>=1.0.0

matplotlib>=3.7.0
plotly>=5.15.0

backtesting>=0.3.3

jupyter>=1.0.0
tqdm>=4.65.0
black
